+++ src/graphics/texture.rs
@@ -1,6 +1,6 @@
 //! Functions and types relating to textures.
 
-use std::cell::RefCell;
+use std::cell::Cell;
 use std::path::Path;
 use std::rc::Rc;
 
@@ -9,6 +9,21 @@ use crate::fs;
 use crate::graphics::{self, DrawParams, Drawable, Rectangle};
 use crate::platform::{GraphicsDevice, RawTexture};
 use crate::Context;
+
+#[derive(Debug)]
+pub(crate) struct TextureSharedData {
+    pub(crate) handle: RawTexture,
+    filter_mode: Cell<FilterMode>,
+}
+
+impl PartialEq for TextureSharedData {
+    fn eq(&self, other: &TextureSharedData) -> bool {
+        // filter_mode should always match what's set on the GPU,
+        // so we can ignore it for equality checks.
+
+        self.handle.eq(&other.handle)
+    }
+}
 
 /// A texture, held in GPU memory.
 ///
@@ -28,7 +43,7 @@ use crate::Context;
 /// and so can be cloned with little overhead.
 #[derive(Debug, Clone, PartialEq)]
 pub struct Texture {
-    pub(crate) handle: Rc<RefCell<RawTexture>>,
+    pub(crate) data: Rc<TextureSharedData>,
 }
 
 impl Texture {
@@ -98,7 +113,13 @@ impl Texture {
     /// * `TetraError::NotEnoughData` will be returned if not enough data is provided to fill
     /// the texture. This is to prevent the graphics API from trying to read uninitialized memory.
     pub fn from_rgba(ctx: &mut Context, width: i32, height: i32, data: &[u8]) -> Result<Texture> {
-        Texture::with_device(&mut ctx.device, width, height, data)
+        Texture::with_device(
+            &mut ctx.device,
+            width,
+            height,
+            data,
+            ctx.graphics.default_filter_mode,
+        )
     }
 
     pub(crate) fn with_device(
@@ -106,6 +127,7 @@ impl Texture {
         width: i32,
         height: i32,
         data: &[u8],
+        filter_mode: FilterMode,
     ) -> Result<Texture> {
         let expected = (width * height * 4) as usize;
         let actual = data.len();
@@ -117,9 +139,13 @@ impl Texture {
         let handle = device.new_texture(width, height)?;
 
         device.set_texture_data(&handle, &data, 0, 0, width, height);
+        device.set_texture_filter_mode(&handle, filter_mode);
 
         Ok(Texture {
-            handle: Rc::new(RefCell::new(handle)),
+            data: Rc::new(TextureSharedData {
+                handle,
+                filter_mode: Cell::new(FilterMode::Linear),
+            }),
         })
     }
 
@@ -127,39 +153,45 @@ impl Texture {
         device: &mut GraphicsDevice,
         width: i32,
         height: i32,
+        filter_mode: FilterMode,
     ) -> Result<Texture> {
         let handle = device.new_texture(width, height)?;
+        device.set_texture_filter_mode(&handle, filter_mode);
 
         Ok(Texture {
-            handle: Rc::new(RefCell::new(handle)),
+            data: Rc::new(TextureSharedData {
+                handle,
+                filter_mode: Cell::new(filter_mode),
+            }),
         })
     }
 
     /// Returns the width of the texture.
     pub fn width(&self) -> i32 {
-        self.handle.borrow().width()
+        self.data.handle.width()
     }
 
     /// Returns the height of the texture.
     pub fn height(&self) -> i32 {
-        self.handle.borrow().height()
+        self.data.handle.height()
     }
 
     /// Returns the size of the canvas.
     pub fn size(&self) -> (i32, i32) {
-        let handle = self.handle.borrow();
-        (handle.width(), handle.height())
+        (self.data.handle.width(), self.data.handle.height())
     }
 
     /// Returns the filter mode being used by the texture.
     pub fn filter_mode(&self) -> FilterMode {
-        self.handle.borrow().filter_mode()
+        self.data.filter_mode.get()
     }
 
     /// Sets the filter mode that should be used by the texture.
     pub fn set_filter_mode(&mut self, ctx: &mut Context, filter_mode: FilterMode) {
         ctx.device
-            .set_texture_filter_mode(&mut *self.handle.borrow_mut(), filter_mode);
+            .set_texture_filter_mode(&self.data.handle, filter_mode);
+
+        self.data.filter_mode.set(filter_mode);
     }
 }