+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file, following t
 
 This project adheres to Semantic Versioning.
 
-## [Upcoming]
+## [0.3.3] - 2020-04-04
 
 ### Added
 
@@ -18,7 +18,7 @@ This project adheres to Semantic Versioning.
 * Updated `image` to 0.23.
 * Updated `rodio` to 0.11.
 
-## [0.3.2] - 2019-01-15
+## [0.3.2] - 2020-01-15
 
 ### Added
 
@@ -441,7 +441,8 @@ This project adheres to Semantic Versioning.
 
 * Initial release!
 
-[Upcoming]: https://github.com/17cupsofcoffee/tetra/compare/0.3.2..HEAD
+[Upcoming]: https://github.com/17cupsofcoffee/tetra/compare/0.3.3..HEAD
+[0.3.3]: https://github.com/17cupsofcoffee/tetra/compare/0.3.2..0.3.3
 [0.3.2]: https://github.com/17cupsofcoffee/tetra/compare/0.3.1..0.3.2
 [0.3.1]: https://github.com/17cupsofcoffee/tetra/compare/0.3.0..0.3.1
 [0.3.0]: https://github.com/17cupsofcoffee/tetra/compare/0.2.20..0.3.0